CBOR: Refactor Tests
